+++ engine/source/unit_tests/MessageManager_test.cpp
@@ -8,13 +8,13 @@ TEST(SW_Engine_MessageManager, mark_as_read)
 
   Messages m = mm.get_unread_messages_and_mark_as_read();
   EXPECT_FALSE(m.empty());
-  EXPECT_EQ(1, m.get_messages().size());
+  EXPECT_EQ(static_cast<uint>(1), m.get_messages().size());
 
   mm.clear_if_necessary();
 
   m = mm.get_unread_messages();
   EXPECT_TRUE(m.empty());
-  EXPECT_EQ(0, m.get_messages().size());
+  EXPECT_EQ(static_cast<uint>(0), m.get_messages().size());
 }
 
 TEST(SW_Engine_MessageManager, add_new_message)
@@ -25,7 +25,7 @@ TEST(SW_Engine_MessageManager, add_new_message)
 
   Messages m = mm.get_unread_messages();
   EXPECT_FALSE(m.empty());
-  EXPECT_EQ(2, m.get_messages().size());
+  EXPECT_EQ(static_cast<uint>(2), m.get_messages().size());
 }
