+++ world/source/unit_tests/Inventory_test.cpp
@@ -73,11 +73,11 @@ TEST(SW_World_Inventory, count_items_with_property)
 
   inv.add(tool);
 
-  EXPECT_EQ(1, inv.count_items_with_property(SmithingConstants::SMITHING_CONSTANTS_MATERIAL_TYPE));
+  EXPECT_EQ(static_cast<uint>(1), inv.count_items_with_property(SmithingConstants::SMITHING_CONSTANTS_MATERIAL_TYPE));
 
   tool->set_quantity(3);
 
-  EXPECT_EQ(3, inv.count_items_with_property(SmithingConstants::SMITHING_CONSTANTS_MATERIAL_TYPE));
+  EXPECT_EQ(static_cast<uint>(3), inv.count_items_with_property(SmithingConstants::SMITHING_CONSTANTS_MATERIAL_TYPE));
 }
 
 TEST(SW_World_Inventory, count_items_without_property)
@@ -90,18 +90,18 @@ TEST(SW_World_Inventory, count_items_without_property)
 
   inv.add(book);
 
-  EXPECT_EQ(1, inv.count_items_without_property(prop));
+  EXPECT_EQ(static_cast<uint>(1), inv.count_items_without_property(prop));
 
   book->set_additional_property(prop, "1");
 
-  EXPECT_EQ(0, inv.count_items_without_property(prop));
+  EXPECT_EQ(static_cast<uint>(0), inv.count_items_without_property(prop));
 }
 
 TEST(SW_World_Inventory, count_items)
 {
   Inventory inv;
 
-  EXPECT_EQ(0, inv.count_items("book"));
+  EXPECT_EQ(static_cast<uint>(0), inv.count_items("book"));
 
   SpellbookPtr book = std::make_shared<Spellbook>();
   book->set_base_id("book");
@@ -110,7 +110,7 @@ TEST(SW_World_Inventory, count_items)
 
   inv.add(book);
 
-  EXPECT_EQ(4, inv.count_items("book"));
+  EXPECT_EQ(static_cast<uint>(4), inv.count_items("book"));
 
   SpellbookPtr book2 = std::make_shared<Spellbook>();
   book2->set_base_id("book");
@@ -119,19 +119,19 @@ TEST(SW_World_Inventory, count_items)
 
   inv.add(book2);
   
-  EXPECT_EQ(8, inv.count_items("book"));
+  EXPECT_EQ(static_cast<uint>(8), inv.count_items("book"));
 }
 
 TEST(SW_World_Inventory, size)
 {
   Inventory inv;
 
-  EXPECT_EQ(0, inv.size());
+  EXPECT_EQ(static_cast<uint>(0), inv.size());
 
   SpellbookPtr book = std::make_shared<Spellbook>();
   inv.add(book);
 
-  EXPECT_EQ(1, inv.size());
+  EXPECT_EQ(static_cast<uint>(1), inv.size());
 }
 
 TEST(SW_World_Inventory, empty)
@@ -159,7 +159,7 @@ TEST(SW_World_Inventory, items_are_persisted)
   inv.add(tool);
 
   EXPECT_FALSE(inv.empty());
-  EXPECT_EQ(2, inv.size());
+  EXPECT_EQ(static_cast<uint>(2), inv.size());
 
   ItemPtr item = inv.get_from_id("book");
   ItemPtr item2 = inv.get_from_id("tool");
@@ -185,7 +185,7 @@ TEST(SW_World_Inventory, add_items)
   list<ItemPtr> items = {book};
   i.set_items(items);
 
-  EXPECT_EQ(1, i.count_items());
+  EXPECT_EQ(static_cast<uint>(1), i.count_items());
 
   ToolPtr tool = std::make_shared<Tool>();
   tool->set_id("tool");
@@ -193,7 +193,7 @@ TEST(SW_World_Inventory, add_items)
   items = {tool};
   i.add_items(items);
 
-  EXPECT_EQ(2, i.count_items());
+  EXPECT_EQ(static_cast<uint>(2), i.count_items());
 }
 
 TEST(SW_World_Inventory, set_additional_property)
@@ -203,7 +203,7 @@ TEST(SW_World_Inventory, set_additional_property)
   ItemPtr item2 = std::make_shared<Spellbook>();
   inv.add_items({ item, item2 });
   inv.set_additional_property(ItemProperties::ITEM_PROPERTIES_BRANDED, "1");
-  EXPECT_EQ(2, inv.count_items_with_property(ItemProperties::ITEM_PROPERTIES_BRANDED));
+  EXPECT_EQ(static_cast<uint>(2), inv.count_items_with_property(ItemProperties::ITEM_PROPERTIES_BRANDED));
 }
 
 TEST(SW_World_Inventory, get_from_type)
@@ -220,7 +220,7 @@ TEST(SW_World_Inventory, get_from_type)
 
   vector<ItemPtr> items = inv.get_from_type(ItemType::ITEM_TYPE_SPELLBOOK);
 
-  EXPECT_EQ(2, items.size());
+  EXPECT_EQ(static_cast<uint>(2), items.size());
 }
 
 TEST(SW_World_Inventory, no_drop_effect_sid)
@@ -247,15 +247,15 @@ TEST(SW_World_Inventory, remove_by_base_id)
   i.add(item2);
   i.add(item3);
 
-  EXPECT_EQ(3, i.size());
+  EXPECT_EQ(static_cast<uint>(3), i.size());
 
   i.remove_by_base_id("book");
 
-  EXPECT_EQ(2, i.size());
+  EXPECT_EQ(static_cast<uint>(2), i.size());
 
   i.remove_by_base_id("book");
 
-  EXPECT_EQ(1, i.size());
+  EXPECT_EQ(static_cast<uint>(1), i.size());
 }
 
 TEST(SW_World_Inventory, remove_by_base_id_with_properties)
@@ -275,18 +275,18 @@ TEST(SW_World_Inventory, remove_by_base_id_with_properties)
   i.add(item2);
   i.add(item3);
 
-  EXPECT_EQ(3, i.size());
+  EXPECT_EQ(static_cast<uint>(3), i.size());
 
   map<string, string> props = { {"prop1", "propval"} };
   i.remove_by_base_id("book", 1, props);
 
-  EXPECT_EQ(3, i.size());
+  EXPECT_EQ(static_cast<uint>(3), i.size());
 
   item->set_additional_properties(props);
 
   i.remove_by_base_id("book", 1, props);
 
-  EXPECT_EQ(2, i.size());
+  EXPECT_EQ(static_cast<uint>(2), i.size());
 }
 
 TEST(SW_World_Inventory, has_item)
@@ -317,7 +317,7 @@ TEST(SW_World_Inventory, get_all_from_base_id)
   i.add(item2);
   i.add(item3);
 
-  EXPECT_EQ(2, i.get_all_from_base_id("book").size());
+  EXPECT_EQ(static_cast<uint>(2), i.get_all_from_base_id("book").size());
 }
 
 TEST(SW_World_Inventory, get_all_from_property)
@@ -335,11 +335,11 @@ TEST(SW_World_Inventory, get_all_from_property)
   i.add(item2);
   i.add(item3);
 
-  EXPECT_EQ(1, i.get_all_from_property(pname).size());
+  EXPECT_EQ(static_cast<uint>(1), i.get_all_from_property(pname).size());
 
   item3->set_additional_property(pname, "test");
 
-  EXPECT_EQ(2, i.get_all_from_property(pname).size());
+  EXPECT_EQ(static_cast<uint>(2), i.get_all_from_property(pname).size());
 }
 
 TEST(SW_World_Inventory, get_all_from_property_and_required_value)
@@ -357,8 +357,8 @@ TEST(SW_World_Inventory, get_all_from_property_and_required_value)
   i.add(item2);
   i.add(item3);
 
-  EXPECT_EQ(0, i.get_all_from_property(pname, "aaa").size());
-  EXPECT_EQ(1, i.get_all_from_property(pname, "test").size());
+  EXPECT_EQ(static_cast<uint>(0), i.get_all_from_property(pname, "aaa").size());
+  EXPECT_EQ(static_cast<uint>(1), i.get_all_from_property(pname, "test").size());
 }
 
 TEST(SW_World_Inventory, merge_or_add)
@@ -380,14 +380,14 @@ TEST(SW_World_Inventory, merge_or_add)
 
   Inventory i;
   EXPECT_TRUE(i.merge_or_add(item3, InventoryAdditionType::INVENTORY_ADDITION_BACK));
-  EXPECT_EQ(1, i.size());
+  EXPECT_EQ(static_cast<uint>(1), i.size());
 
   EXPECT_TRUE(i.merge_or_add(item2, InventoryAdditionType::INVENTORY_ADDITION_BACK));
-  EXPECT_EQ(2, i.size());
+  EXPECT_EQ(static_cast<uint>(2), i.size());
 
   EXPECT_TRUE(i.merge_or_add(item, InventoryAdditionType::INVENTORY_ADDITION_BACK));
-  EXPECT_EQ(2, i.size());
-  EXPECT_EQ(2, i.get_from_base_id("abc212")->get_quantity());
+  EXPECT_EQ(static_cast<uint>(2), i.size());
+  EXPECT_EQ(static_cast<uint>(2), i.get_from_base_id("abc212")->get_quantity());
 }
 
 TEST(SW_World_Inventory, merge_or_add_whole_inventory)
@@ -403,8 +403,8 @@ TEST(SW_World_Inventory, merge_or_add_whole_inventory)
   inv2->merge_or_add(inv, InventoryAdditionType::INVENTORY_ADDITION_BACK);
 
   EXPECT_FALSE(inv->merge_or_add(inv, InventoryAdditionType::INVENTORY_ADDITION_FRONT));
-  EXPECT_EQ(1, inv->size());
-  EXPECT_EQ(1, inv2->size());
+  EXPECT_EQ(static_cast<uint>(1), inv->size());
+  EXPECT_EQ(static_cast<uint>(1), inv2->size());
 }
 
 TEST(SW_World_Inventory, transfer_to)
@@ -420,7 +420,7 @@ TEST(SW_World_Inventory, transfer_to)
   inv->transfer_to(inv2);
 
   EXPECT_FALSE(inv->transfer_to(inv));
-  EXPECT_EQ(0, inv->size());
-  EXPECT_EQ(1, inv2->size());
+  EXPECT_EQ(static_cast<uint>(0), inv->size());
+  EXPECT_EQ(static_cast<uint>(1), inv2->size());
 }
